add field on course dao,service and controller

+++ features/src/main/java/com/lawencon/elearning/dto/teacher/TeacherReportResponseDTO.java
@@ -14,5 +14,5 @@ public class TeacherReportResponseDTO {
   private Integer totalUjian;
   private Integer totalAssignment;
   private Integer notAssignment;
-  private Integer avgScore;
+  private Double avgScore;
 }
